requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.1
dj-rest-auth==3.0.0
Django==4.1.6
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
jwt==1.3.1
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.3
typing==3.7.4.3
tzdata==2023.3
tzlocal==4.3
whitenoise==6.4.0
